nat: GC NAT entries created by DSR
In the case of DSR, the following CT and NAT entries are created on
a host which runs a service endpoint and to which a client request is
forwarded:

* NAT: endpoint -> client (XLATE_SRC aka TUPLE_F_OUT)
* CT:  client -> endpoint (TUPLE_F_IN)

Previously, the CT GC ignored NAT entries when a corresponding CT entry was
of the TUPLE_F_IN type. Therefore, the DSR NAT entries could not have been
collected.

Signed-off-by: Martynas Pumputis <m@lambda.lt>